pyproject.toml
[project]
name = "flafka-api-spotipy"
version = "0.1.0"
description = ""
authors = [
    {name = "Guilherme Belisario",email = "gbo2000@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "spotipy (>=2.25.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pytest (>=8.3.4,<9.0.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
